@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -54,6 +54,7 @@ flutter:
     - assets/opensans/
     - assets/lang/ar.json
     - assets/lang/en.json
+    - assets/data/database.db
 
   fonts:
     - family: cairo
