@@ -26,6 +26,7 @@ class DetailPage extends StatefulWidget {
 }
 
 class _DetailPageState extends State<DetailPage> with TickerProviderStateMixin  {
+
   double percentComplete;
   AnimationController animationBar;
   double barPercent = 0.0;
@@ -197,7 +198,7 @@ class _DetailPageState extends State<DetailPage> with TickerProviderStateMixin  
                         menu.tasks.length > 0 ? getIconText(
                           icon: Icon(
                             isDelete ? Icons.check : Icons.edit,
-                            color: menu.themeMenu.color,
+                            color: menu.theme.color,
                           ),
                           text: isDelete ? getTranslated(context, "back") : getTranslated(context, "edit"),
                           onTap:() {
@@ -210,22 +211,22 @@ class _DetailPageState extends State<DetailPage> with TickerProviderStateMixin  
                         Padding(
                           padding: EdgeInsets.only(left: 20.0, right: 20.0),
                           child: new CircleAvatar(
-                            backgroundColor: menu.themeMenu.color,
-                            radius: 40,
+                            backgroundColor: menu.theme.color,
+                            radius: 24,
                             child: Hero(
                               tag: menu.id + "_icon",
                               child: Container(
                                 decoration: BoxDecoration(
                                   shape: BoxShape.circle,
-                                  color: menu.themeMenu.color,
+                                  color: menu.theme.color,
                                   border: Border.all(color: Colors.grey.withAlpha(70), style: BorderStyle.solid, width: 1.0),
                                 ),
                                 child: Padding(
                                   padding: EdgeInsets.all(8.0),
                                   child: Icon(
-                                    menu.themeMenu.icon,
+                                    menu.theme.icon,
                                     color: Colors.white,
-                                    size: 40,
+                                    size: 24,
                                   ),
                                 ),
                               ),
@@ -249,7 +250,7 @@ class _DetailPageState extends State<DetailPage> with TickerProviderStateMixin  
                                 child: LinearProgressIndicator(
                                   value: barPercent,
                                   backgroundColor: Colors.grey.withAlpha(50),
-                                  valueColor: AlwaysStoppedAnimation<Color>(menu.themeMenu.color),
+                                  valueColor: AlwaysStoppedAnimation<Color>(menu.theme.color),
                                 ),
                               ),
                               Padding(
@@ -315,8 +316,8 @@ class _DetailPageState extends State<DetailPage> with TickerProviderStateMixin  
 
                                 },
                                 child: Icon(
-                                  menu.themeMenu.icon,
-                                  color: menu.themeMenu.color,
+                                  menu.theme.icon,
+                                  color: menu.theme.color,
                                 ),
                               ),
                               suffixIcon: GestureDetector(
@@ -330,14 +331,14 @@ class _DetailPageState extends State<DetailPage> with TickerProviderStateMixin  
                                   decoration: BoxDecoration(
                                       borderRadius: BorderRadius.circular(8),
                                       color: completed()
-                                          ? menu.themeMenu.color
+                                          ? menu.theme.color
                                           : Colors.grey.withOpacity(0.2),
                                   ),
                                   child: Icon(
                                     Icons.check,
                                     color: completed()
                                         ? Colors.black
-                                        : menu.themeMenu.color,
+                                        : menu.theme.color,
                                     size: 20,
                                   ),
                                 ),
@@ -352,7 +353,7 @@ class _DetailPageState extends State<DetailPage> with TickerProviderStateMixin  
                               getIconText(
                                 icon: Icon(
                                   Icons.timer,
-                                  color: menu.themeMenu.color,
+                                  color: menu.theme.color,
                                 ),
                                 text: _taskObject.startDate == 0 ? getTranslated(context, "startingTime") : getMissionTime(timestamp: _taskObject.startDate),
                                 onTap:() async{
@@ -417,7 +418,7 @@ class _DetailPageState extends State<DetailPage> with TickerProviderStateMixin  
     switch(isDelete){
       case false:
         return new CustomCheckboxListTile(
-          activeColor: menu.themeMenu.color,
+          activeColor: menu.theme.color,
           value: task.isCompleted,
           onChanged: (value) {
             _updateTaskObject(task: task);
@@ -450,7 +451,7 @@ class _DetailPageState extends State<DetailPage> with TickerProviderStateMixin  
         break;
       default:
         return new CustomCheckboxListTile(
-          activeColor: menu.themeMenu.color,
+          activeColor: menu.theme.color,
           value: task.isCompleted,
           onChanged: (value) {
             _updateTaskObject(task: task);
